pyproject.toml
[tool.poetry]
name = "asr_standardized_combined"
version = "1.0.0"
description = "Assembles and cleans Norwegian speech corpora"
authors = ["Per Erik Solberg <per.solberg@nb.no>", "Pablo Ortiz <pablo.ortiz@telenor.com>", "Phoebe Parsons <phoebe.l.t.parsons@ntnu.no>", "Torbjørn Svendsen <torbjorn.svendsen@ntnu.no>", "Giampiero Salvi <giampiero.salvi@ntnu.no>"]
readme = "README.md"
license = "Apache-2.0"
repository = "https://github.com/scribe-project/asr-standardized-combined"
include = ["LICENCE"]

[tool.poetry.dependencies]
python = "^3.6.2"
beautifulsoup4 = "4.11.1"
pandas = "1.4.0"
numpy = "1.21.0"
lxml = "4.9.0"
pydub = "0.25.1"
pympi-ling = "1.70.2"
tqdm = "4.63.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"